+++ tests/test_dispatcher_cli.py
