+++ tests/test_hooks.py
@@ -17,9 +17,9 @@ import repokid.cli.repokid_cli
 import repokid.hooks
 from repokid.hooks.loggers import log_during_repoable_calculation_batch_hooks
 from repokid.role import Role
-from repokid.tests.artifacts.hook import function_1
-from repokid.tests.artifacts.hook import function_2
-from repokid.tests.test_commands import ROLES
+from tests.artifacts.hook import function_1
+from tests.artifacts.hook import function_2
+from tests.test_commands import ROLES
 
 
 def func_a(input_dict):
