+++ tests/vars.py
